requirements.txt
absl-py==0.9.0
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
deepdiff==5.0.1
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
enum34==1.1.10
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
higher==0.2.1
idna==2.10
imageio==2.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.2
jsonpickle==1.4.1
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
kiwisolver==1.2.0
llvmlite==0.34.0
Markdown==3.2.2
matplotlib==3.2.2
mock==4.0.2
more-itertools==8.4.0
mpyq==0.2.5
munch==2.5.0
nbformat==5.0.7
notebook==6.0.3
numba==0.51.2
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
ordered-set==4.0.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathlib2==2.3.5
pbr==5.4.5
Pillow==7.2.0
pip-review==1.1.0
pluggy==0.13.1
portpicker==1.3.1
probscale==0.2.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
py==1.9.0
py-cpuinfo==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySC2==3.0.0
pytest==5.4.3
python-dateutil==2.8.1
python-ternary==1.0.7
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s2clientprotocol==4.11.4.78285.0
s2protocol==4.11.4.78285.0
sacred==0.8.1
scipy==1.5.0
Send2Trash==1.5.0
six==1.15.0
sk-video==1.1.10
smmap==3.0.4
snakeviz==2.1.0
tensorboard==2.2.2
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
# torch==1.5.1 # for normal
# https://pypi.virtaitech.com torch==1.5.0+cu101 # pypi.virtaitech.com
torchvision==0.6.0 # 0.6.1 for torch==1.6.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
pyclustering==0.10.1.2